Project Restarted
Scrapped a bunch of the old stuff and started fresh. Very Basic "Singleplayer" Bingo interface is finished with Left and Mouse buttons toggling blue or orange completion. A JSON file can be read with tasks but currently the DESC has no function. Pushing this change because I will be updating to the newest verison of raylib and don't want to break anything. This was working with Raylib 3.0.